requirements.txt
asgiref==3.5.2
boto3==1.23.10
botocore==1.26.10
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==37.0.2
Django==4.0.4
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-ckeditor==6.4.2
django-cryptography==1.1
django-environ==0.9.0
django-js-asset==2.0.0
django-login-required-middleware==0.8.0
django-storages==1.12.3
django-widget-tweaks==1.4.12
idna==3.3
jmespath==1.0.0
mailchimp-marketing==3.0.75
numpy==1.22.4
pandas==1.4.3
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
stripe==3.5.0
tzdata==2022.1
urllib3==1.26.9
